+++ projects/brochure-marketplace/src/app/services/live-api.service.ts
@@ -107,9 +107,8 @@ export class LiveApiService extends ApiService {
 
   private async httpRequest<T>(opts: HttpOptions): Promise<T> {
     const res = await this.http.httpRequest<T>(opts)
-    if (res.headers.get('Repr-Digest')) {
-      // verify
-      const digest = res.headers.get('Repr-Digest')!
+    if (res.headers.get('File-Digest')) {
+      const digest = res.headers.get('File-Digest')!
       let data: Uint8Array
       if (opts.responseType === 'arrayBuffer') {
         data = Buffer.from(res.body as ArrayBuffer)
@@ -119,7 +118,7 @@ export class LiveApiService extends ApiService {
         data = Buffer.from(await (res.body as Blob).arrayBuffer())
       } else {
         console.warn(
-          `could not verify Repr-Digest for responseType ${
+          `could not verify File-Digest for responseType ${
             opts.responseType || 'json'
           }`,
         )