+++ projects/start-os/web/ui/src/app/components/refresh-alert.component.ts
@@ -2,11 +2,10 @@ import { Component, inject } from '@angular/core'
 import { toSignal } from '@angular/core/rxjs-interop'
 import { SwUpdate } from '@angular/service-worker'
 import { WA_WINDOW } from '@ng-web-apis/common'
-import { i18nPipe } from '@start9labs/shared'
+import { i18nPipe, TaskService } from '@start9labs/shared'
 import { Version } from '@start9labs/start-core'
 import { TuiResponsiveDialog } from '@taiga-ui/addon-mobile'
 import { TuiButton } from '@taiga-ui/core'
-import { TuiNotificationMiddleService } from '@taiga-ui/kit'
 import { PatchDB } from 'patch-db-client'
 import { distinctUntilChanged, map, merge, Subject } from 'rxjs'
 import { ConfigService } from 'src/app/services/config.service'
@@ -30,45 +29,23 @@ import { DataModel } from 'src/app/services/patch-db/data-model'
               | i18n
           }}
         </p>
-        <button
-          tuiButton
-          appearance="secondary"
-          style="float: right"
-          [tuiAppearanceFocus]="false"
-          (click)="pwaReload()"
-        >
-          {{ 'Refresh' | i18n }}
-        </button>
       } @else {
-        {{
-          'Your user interface is cached and out of date. Hard refresh the page to get the latest UI.'
-            | i18n
-        }}
-        <ul>
-          <li>
-            <b>On Mac (Chrome/Firefox)</b>
-            : cmd + shift + R
-          </li>
-          <li>
-            <b>On Mac (Safari)</b>
-            : option + cmd + R, or hold option and choose View > Reload Page
-            from Origin
-          </li>
-          <li>
-            <b>On Linux/Windows</b>
-            : ctrl + shift + R
-          </li>
-        </ul>
-        <button
-          tuiButton
-          appearance="secondary"
-          style="float: right"
-          [tuiAppearanceFocus]="false"
-          (click)="dismiss$.next()"
-        >
-          {{ 'Ok' | i18n }}
-        </button>
+        <p>
+          {{
+            'StartOS has been updated, but this page is still running the previous interface. Refresh the page to get the latest version.'
+              | i18n
+          }}
+        </p>
       }
+      <button
+        tuiButton
+        appearance="secondary"
+        style="float: right"
+        [tuiAppearanceFocus]="false"
+        (click)="reload()"
+      >
+        {{ 'Refresh' | i18n }}
+      </button>
     </ng-template>
   `,
   imports: [TuiResponsiveDialog, TuiButton, i18nPipe],
@@ -76,7 +53,7 @@ import { DataModel } from 'src/app/services/patch-db/data-model'
 export class RefreshAlertComponent {
   private readonly win = inject(WA_WINDOW)
   private readonly updates = inject(SwUpdate)
-  private readonly loader = inject(TuiNotificationMiddleService)
+  private readonly tasks = inject(TaskService)
   private readonly version = Version.parse(inject(ConfigService).version)
 
   readonly i18n = inject(i18nPipe)
@@ -99,15 +76,18 @@ export class RefreshAlertComponent {
     },
   )
 
-  async pwaReload() {
+  protected async reload(): Promise<void> {
     try {
-      this.loader.open('Reloading PWA').subscribe()
-      // attempt to update to the latest client version available
-      await this.updates.activateUpdate()
-    } catch (e) {
-      console.error('Error activating update from service worker: ', e)
+      if (
+        this.updates.isEnabled &&
+        this.win.navigator.serviceWorker.controller !== null
+      ) {
+        await this.tasks.run(async () => {
+          await this.updates.checkForUpdate()
+          await this.updates.activateUpdate()
+        }, 'Loading')
+      }
     } finally {
-      // always reload, as this resolves most out of sync cases
       this.win.location.reload()
     }
   }